improve backtest average price accuracy

+++ ff_standard_lib/src/messages/data_server_messaging.rs
@@ -234,7 +234,7 @@ DataServerResponse {
 
     DecimalAccuracy{
         callback_id: u64,
-        accuracy: u8
+        accuracy: u32
     },
 
     ValuePerTick{
